+++ runtime/spark-tests/src/intTest/java/org/apache/polaris/service/spark/it/CatalogFederationIT.java
@@ -41,8 +41,8 @@ public class CatalogFederationIT extends CatalogFederationIntegrationTest {
           .put("polaris.features.\"SKIP_CREDENTIAL_SUBSCOPING_INDIRECTION\"", "false")
           .put("polaris.features.\"ALLOW_EXTERNAL_CATALOG_CREDENTIAL_VENDING\"", "true")
           .put("polaris.features.\"ALLOW_FEDERATED_CATALOGS_CREDENTIAL_VENDING\"", "true")
-          .put("polaris.storage.aws.access-key", CatalogFederationIntegrationTest.MINIO_ACCESS_KEY)
-          .put("polaris.storage.aws.secret-key", CatalogFederationIntegrationTest.MINIO_SECRET_KEY)
+          .put("polaris.storage.aws.access-key", CatalogFederationIntegrationTest.RUSTFS_ACCESS_KEY)
+          .put("polaris.storage.aws.secret-key", CatalogFederationIntegrationTest.RUSTFS_SECRET_KEY)
           .build();
     }
   }